requirements.txt
rich>=13.0.0
plotly>=5.18.0
pyyaml>=6.0
pandas>=2.0.0
